requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-countries==5.3.3
django-crispy-forms==1.7.2
django-elasticsearch-dsl==0.5.1
django-extensions==2.1.9
django-import-export==1.2.0
elasticsearch==6.4.0
elasticsearch-dsl==6.1.0
et-xmlfile==1.0.1
idna==2.8
ipaddress==1.0.22
jdcal==1.4.1
oauthlib==3.0.2
odfpy==1.4.0
openpyxl==2.6.2
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
